updates for milestone3, classes added to main.rb file
