+++ bin/users.rb
@@ -1,0 +1,26 @@
+# rubocop:disable Metrics/PerceivedComplexity,Metrics/CyclomaticComplexity
+
+class Users
+  attr_accessor :player1, :player2, :next_player
+  attr_accessor :first_player_selections, :second_player_selections
+
+  def initialize(player1, player2)
+    @player1 = player1
+    @player2 = player2
+    @next_player = false
+    @first_player_selections = []
+    @second_player_selections = []
+  end
+
+  def winner
+    player_name = @next_player ? @player2 : @player1
+    puts "#{player_name} has won the game!"
+  end
+
+  def draw
+    player_name = @next_player ? @player2 : @player1
+    puts "It's a tie, neither #{player1} nor #{player2} was able to win the game!"
+  end
+end
+
+# rubocop:enable Metrics/PerceivedComplexity,Metrics/CyclomaticComplexity
